[feat] 앙파상 룰 강화

+++ src/main/java/chess/model/service/ChessGameService.java
@@ -8,6 +8,7 @@ import chess.model.domain.board.ChessGame;
 import chess.model.domain.board.EnPassant;
 import chess.model.domain.board.TeamScore;
 import chess.model.domain.piece.Color;
+import chess.model.domain.piece.Pawn;
 import chess.model.domain.piece.Piece;
 import chess.model.domain.piece.Type;
 import chess.model.domain.state.MoveOrder;
@@ -93,16 +94,19 @@ public class ChessGameService {
         ChessGame chessGame = getChessGame(gameId);
         boolean canCastling = chessGame.canCastling(moveSquare);
         boolean pawnSpecialMove = chessGame.isPawnSpecialMove(moveSquare);
+        boolean movePawn = chessGame.whoMovePiece(moveSquare) instanceof Pawn;
         MoveState moveState = chessGame.movePieceWhenCanMove(moveSquare);
         Color gameTurn = CHESS_GAME_DAO.getGameTurn(gameId).orElseThrow(IllegalAccessError::new);
         if (moveState.isSucceed()) {
+            CHESS_BOARD_DAO.deleteMyEnpassant(gameId);
             CHESS_BOARD_DAO.deleteBoardSquare(gameId, moveSquare.get(MoveOrder.AFTER));
             CHESS_BOARD_DAO.copyBoardSquare(gameId, moveSquare);
             CHESS_BOARD_DAO.deleteBoardSquare(gameId, moveSquare.get(MoveOrder.BEFORE));
             if (pawnSpecialMove) {
                 CHESS_BOARD_DAO.updateEnPassant(gameId, moveSquare);
             }
-            if (enPassant.hasOtherEnpassant(moveSquare.get(MoveOrder.AFTER), gameTurn)) {
+            if (movePawn && enPassant
+                .hasOtherEnpassant(moveSquare.get(MoveOrder.AFTER), gameTurn)) {
                 CHESS_BOARD_DAO.deleteEnpassant(gameId, moveSquare.get(MoveOrder.AFTER));
             }
             if (canCastling) {
